Fully working project withoout extra features

+++ app/Http/Controllers/AdminMediasController.php
@@ -100,7 +100,9 @@ class AdminMediasController extends Controller
         $photo = Photo::findOrFail($id);
 
 
-        unlink(public_path().$photo->file);
+        if(file_exists(public_path() . $photo->file)){
+            unlink(public_path() . $photo->file);
+        }
 
         $photo->delete();
         $users = User::all();
